Update README.md
Guat-ED

The Problem:
Education Quality. The causes of Guatemala’s larger gaps are likely related to larger differences across indigenous and nonindigenous students in variables that affect test scores, such as family income or school resources.

The Solution
Providing quality education in all rural areas of Guatemala is a big challenge.
However, technology can be utilized to improve teaching and learning in rural communities.

An application will be created to identify schools that need to improve their educational standards ( based on the academic performance of their students).

Guat-ED will help organizations or individuals identify which schools need the most help.